# integrations/financial-institution-agent/openapi.yaml
openapi: 3.0.3
info:
  title: Canonical Banking Gateway for Dify Agents
  version: 0.1.0
  x-dify-identity:
    mode: forward-end-user-identity
    header: X-Dify-End-User-ID
    note: Enable for dynamic end-user context. Private data additionally requires the institution proxy.
  x-institution-security:
    private_data_requires: external-identity-proxy
    verified_subject_header: X-Verified-Subject
    verified_session_state_header: X-Verified-Session-State
    note: The proxy injects these trusted headers; the model and user must never provide them.
  description: |
    Core-agnostic contract for customer-facing financial institution agents.
    The gateway behind this contract may use any banking core. Dify must not
    connect directly to a core or its database.
servers:
  - url: https://banking-gateway.example.internal
    description: Institution private gateway. Replace per deployment.
security:
  - bearerAuth: []
...
paths:
...
  /v1/me:
    get:
      tags: [Customers]
      operationId: getCurrentCustomer
      summary: Get the authenticated customer's profile
      description: Resolves the customer from the trusted session; never ask the user for a customerId.
      responses:
        '200':
          description: Authenticated customer profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Customer' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/me/accounts:
    get:
      tags: [Accounts]
      operationId: listCurrentCustomerAccounts
      summary: List accounts for the authenticated customer
      responses:
        '200':
          description: Accounts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Account' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/me/transactions:
    get:
      tags: [Accounts]
      operationId: listCurrentCustomerTransactions
      summary: List transactions for the authenticated customer
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 10, default: 10 }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Transaction' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/me/loans:
    get:
      tags: [Loans]
      operationId: listCurrentCustomerLoans
      summary: List loans for the authenticated customer
      parameters:
        - name: status
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Loans
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Loan' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/me/cards:
    get:
      tags: [Accounts]
      operationId: listCurrentCustomerCards
      summary: List cards for the authenticated customer
      responses:
        '200':
          description: Cards
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Card' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/me/certificates:
    get:
      tags: [Accounts]
      operationId: listCurrentCustomerCertificates
      summary: List certificates for the authenticated customer
      responses:
        '200':
          description: Certificates
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Certificate' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/me/beneficiaries:
    get:
      tags: [Payments]
      operationId: listCurrentCustomerBeneficiaries
      summary: List beneficiaries for the authenticated customer
      responses:
        '200':
          description: Beneficiaries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Beneficiary' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/customers/{customerId}:
    get:
      tags: [Customers]
      operationId: getCustomer
      summary: Get an authorized customer profile
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      responses:
        '200':
          description: Customer profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Customer' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/customers/{customerId}/accounts:
    get:
      tags: [Accounts]
      operationId: listCustomerAccounts
      summary: List authorized customer accounts
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      responses:
        '200':
          description: Accounts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Account' }
  /v1/customers/{customerId}/transactions:
    get:
      tags: [Accounts]
      operationId: listCustomerTransactions
      summary: List recent authorized customer transactions
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - name: limit
          in: query
          description: Keep conversational responses short. Maximum 10 for this template.
          schema: { type: integer, minimum: 1, maximum: 10, default: 10 }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Transaction' }
  /v1/customers/{customerId}/loans:
    get:
      tags: [Loans]
      operationId: listCustomerLoans
      summary: List authorized customer loans
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - name: status
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Loans
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Loan' }
  /v1/accounts/{accountId}:
    get:
      tags: [Accounts]
      operationId: getAccount
      summary: Get an authorized account
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        '200':
          description: Account
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Account' }
  /v1/customers/{customerId}/cards:
    get:
      tags: [Accounts]
      operationId: listCustomerCards
      summary: List authorized customer cards
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      responses:
        '200':
          description: Cards
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Card' }
  /v1/customers/{customerId}/certificates:
    get:
      tags: [Accounts]
      operationId: listCustomerCertificates
      summary: List authorized account certificates
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      responses:
        '200':
          description: Certificates
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Certificate' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/customers/{customerId}/beneficiaries:
    get:
      tags: [Payments]
      operationId: listCustomerBeneficiaries
      summary: List authorized beneficiaries
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      responses:
        '200':
          description: Beneficiaries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Beneficiary' }
  /v1/cards/{cardId}/transactions:
    get:
      tags: [Accounts]
      operationId: listCardTransactions
      summary: List recent transactions for an authorized card
      parameters:
        - $ref: '#/components/parameters/CardId'
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 10, default: 10 }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Card transactions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Transaction' }
  /v1/loans/{loanId}:
    get:
      tags: [Loans]
      operationId: getLoan
      summary: Get an authorized loan
      parameters:
        - $ref: '#/components/parameters/LoanId'
      responses:
        '200':
          description: Loan detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Loan' }
  /v1/loans/{loanId}/schedule:
    get:
      tags: [Loans]
      operationId: getLoanSchedule
      summary: Get the repayment schedule for an authorized loan
      parameters:
        - $ref: '#/components/parameters/LoanId'
      responses:
        '200':
          description: Repayment schedule
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Installment' }
...
  /v1/operations/{operationId}:
    get:
      tags: [Applications]
      operationId: getOperationStatus
      summary: Get the status of an authorized pending operation
      parameters:
        - name: operationId
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 128 }
      responses:
        '200':
          description: Operation status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OperationStatus' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Short-lived gateway token. In dify-user mode, Dify also forwards the end-user identity header.
...
  schemas:
...
    Customer:
      type: object
      required: [id, displayName, status]
      properties:
        id: { type: string }
        displayName: { type: string }
        status: { type: string }
        maskedDocument: { type: string }
        maskedMobile: { type: string }
        branchName: { type: string }
...
    Beneficiary:
      type: object
      required: [id, name, type, status]
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string }
        maskedAccount: { type: string }
        status: { type: string }
...
  responses:
...
    Forbidden:
      description: The session is not authorized for this resource
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Error' }
...